+++ packages/sources/src/components/info-button/info-button.args.ts
@@ -4,6 +4,7 @@ import { InfoButton } from './info-button.models';
 
 export interface InfoButtonArgs {
   active?: boolean;
+  secondary?: boolean;
   label: string;
   click: HandlerFunction;
 }
@@ -19,6 +20,11 @@ export const infoButtonArgTypes: ArgTypes<InfoButtonArgs> = {
       type: 'boolean'
     }
   },
+  secondary: {
+    control: {
+      type: 'boolean'
+    }
+  },
   click: {
     action: 'toggled'
   }
@@ -28,6 +34,7 @@ export function infoButtonArgsMapper(a: InfoButtonArgs): InfoButton {
   return {
     onClick: e => a.click(e),
     active: a.active,
+    secondary: a.secondary,
     label: a.label
   };
 }