+++ packages/sources/src/components/info-button/info-button.models.ts
@@ -1,5 +1,6 @@
 export interface InfoButton {
   active?: boolean;
+  secondary?: boolean;
   label?: string;
   onClick: (e: MouseEvent) => void;
 }